backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9
authlib==1.3.2
httpx==0.27.2
pypdf==4.3.1
python-dotenv==1.0.1
